+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Categories and categoriesIds to compatibility layer from all trees.
+
 ## [1.14.1] - 2020-08-06
 
 ### Fixed